Moves config to TOML and /config.toml (#33)
This commit moves the config file format from .ini to .toml, since it
was parsed as TOML already anywa. The big difference however is that in
the config file, `override_zipcodes` is no longer a comma-separated
string, but has become an actual array.

So we went from:

`override_zipcodes = "88488,90201"`

to

`override_zipcodes = ["88488", "90210"]`

The main reason for this is that I added the `/config` path to
display the running configuration in the actual configuraion format.
However, by serializing the config as TOML, the comma-separated string
became an array. I realized that the config file wasn't proper .ini
anyway, since .ini comments with `;`, rather than '#', which confuses a
bunch of editors.

So, here we standardize to actual TOML. Apart from the implementation,
this commit also updates documentation and refers to `config`
rather than `config.ini` in other places (like startup scripts and
Docker configuration).

However, old configurations will still work. I built in backwards
compatibility for comma-separated config and old paths will still work.
