--- assets/config.ini.example
@@ -1,68 +1,0 @@
-### Mandatory configuration
-username = "my_locast@username"
-password = "very_secret_password"
-
-### Optional configuration
-
-## Address of the interface to bind to. The default value is 127.0.0.1 (localhost). To bind to all
-## interfaces, use 0.0.0.0
-# bind_address = "127.0.0.1"
-
-## What zip codes to override. By default this is not set and locast2tuner will use your external IP
-## to determine your location.
-# override_zipcodes="98101,94102,90210,85006,80123,57701,57103,51054,55111,75001,77002,60601,46077,30301,33602,33125,33403,20002,21206,19103,18505,10009,02115"
-
-## How many days of EPG data to fetch. The default is 8, which is the maximum of data locast.org provides
-# days = 8
-
-## Verbosity. 0 = Info, 1 = Info + HTTP request lgos, 2 = Debug, 3 = Trace
-# verbose = 1
-
-## TCP port to bind to. The default is 6077. In case you override muliple zip codes, locast2tuner will
-## bind to multiple ports, starting at the number specified below (or default 6077). Any additional
-## instance will bind to a port incremented by 1. E.g. if you override 3 zip codes, 3 instances will
-## be started and bound to 6077, 6078 and 6079. In order to only use one instance, use "multiplex"
-## (see below).
-# port = 6077
-
-## The amount of tuners that is communicated to Plex. This will limit the amount of streams that Plex will
-## allow. Note that this is not a limitation in locast2tuner.
-# tuner_count = 3
-
-## Disable caching of station information. By default locast2tuner caches station information for an hour
-## (see cache_timeout below). By disabling the cache, every request for station information will lead
-## to a call to locast.org. Normally you shouldn't have to disable the cache.
-# disable_station_cache = false
-
-## How often (in seconds) the station cache is refreshed.
-# cache_timeout = 3600
-
-## Where to save station data
-# cache_dir = "/var/lib/locast2tuner" # default: $HOME/.locast2tuner
-
-## Normally, when you override muliple zip codes, locast2tuner starts multiple instances (see "bind_address"),
-## but with "multiplex = true", stations from multiple locations will be available through a single instance.
-# multiplex = false
-
-## Remap channel numbers when multiplexing. In case you override multiple zip codes, Emby and Plex will sort
-## channels by channel number, which means channels from different locations might be intermingled. In order
-## circumvent this, you can use "remap = true". This causes locast2tuner to rewrite the channel number based
-## on the amount of instances there are. Locast will remap a "channel_number" to
-## "channel_number + 100 * instance_number", where the instance_number starts at 0.
-## E.g. you override 3 zip codes, then the channels from the first location will be untouched (since 100*0 == 0),
-## the stations for the second location will start at 100 (e.g. 2.1 CBS becomes 102.1 CBS) and the stations
-## for the third location will start at 200 (e.g. 13.2 WWFF becomes 213.2 WWFF).
-# remap = false
-
-## Start the SSDP server in order for Plex or Emby to auto discover the tuner. Currently, the SSDP
-## server isn't working.
-# ssdp = false
-
-## Device firmware that is reported to Plex or Emby. Defaults to "homerun3_atsc"
-# device_firmware = "homerun3_atsc"
-
-## Device model that is reported to Plex or Emby. Defaults to "HDHR3-US"
-# device_model = "HDHR3-US"
-
-## Device version that is reported to Plex or Emby. Defaults to "20170612"
-# device_version = "20170612"
