+++ assets/docker/Dockerfile
@@ -1,4 +1,4 @@
 FROM alpine:latest
 
 COPY ./amd64binaries/locast2tuner /
-ENTRYPOINT ["/locast2tuner", "--config", "/app/config/config.ini", "-b", "0.0.0.0"]
+ENTRYPOINT ["/locast2tuner", "--config", "/app/config/config", "-b", "0.0.0.0"]
